requirements.txt
cx_Oracle
pandas
google-cloud-bigquery